+++ cucumber-compatibility-kit/javascript/features/attachments/attachments.ts
@@ -1,8 +1,17 @@
 import { When } from 'fake-cucumber'
 
-When('the string {word} is attached as {word}', function(
+When('the string {string} is attached as {string}', function(
   text: string,
-  contentType: string
+  mediaType: string
 ) {
-  this.attach(text, contentType)
+  this.attach(text, mediaType)
+})
+
+When('{int} bytes are attached as {string}', function(
+  size: number,
+  mediaType: string
+) {
+  const data = [...Array(size).keys()]
+  const buffer = Buffer.from(data)
+  this.attach(buffer, mediaType)
 })