+++ fake-cucumber/javascript/src/IWorld.ts
@@ -1,5 +1,7 @@
+import { Readable } from 'stream'
+
 export default interface IWorld {
   testStepId: string
 
-  attach(data: string, contentType: string): void
+  attach(data: string | Buffer | Readable, mediaType: string): void
 }