+++ fake-cucumber/javascript/src/TestCase.ts
@@ -3,6 +3,7 @@ import { MessageNotifier } from './types'
 import { messages, TimeConversion } from 'cucumber-messages'
 import { performance } from 'perf_hooks'
 import IWorld from './IWorld'
+import { Readable } from 'stream'
 
 const { millisecondsToDuration } = TimeConversion
 
@@ -10,7 +11,10 @@ class DefaultWorld implements IWorld {
   public testStepId: string
 
   constructor(
-    public readonly attach: (data: any, contentType: string) => void
+    public readonly attach: (
+      data: string | Buffer | Readable,
+      mediaType: string
+    ) => void
   ) {}
 }
 
@@ -48,22 +52,27 @@ export default class TestCase {
       })
     )
 
-    function attach(data: string, contentType: string) {
+    function attach(data: string | Buffer, mediaType: string) {
       if (!this.testStepId) {
         throw new Error(`this.testStepId is not set`)
       }
-      const encoding = messages.Media.Encoding.UTF8 // TODO: Use Base64 is the data is a Buffer (objects will be JSONified)
+      const attachment = new messages.Attachment({
+        testCaseStartedId,
+        testStepId: this.testStepId,
+        mediaType,
+      })
+
+      if (typeof data === 'string') {
+        attachment.text = data
+      } else if (Buffer.isBuffer(data)) {
+        attachment.binary = data
+      } else {
+        throw new Error(`data must be string or Buffer`)
+      }
+
       notifier(
         new messages.Envelope({
-          attachment: new messages.Attachment({
-            data,
-            testCaseStartedId,
-            testStepId: this.testStepId,
-            media: new messages.Media({
-              contentType,
-              encoding,
-            }),
-          }),
+          attachment,
         })
       )
     }